cobuild: handle fee payment and set change output
I call payFee before filling the witness in the previous version. The fee estimation is smaller because the tx size is smaller than the final one after adding all the witnesses. I set feeRate to 3000 to work around this.

In this commit, I refactor code to make it possible to create a dummy tx for fee estimation. It's guaranteed that the dummy tx is not less than the final tx.